コンポーネントの分離
